create separate folders for tests
