+++ .gitignore
@@ -10,7 +10,7 @@ npm-debug.*
 web-build/
 web-report/
 coverage/**/*
-__tests__/__snapshots__
+__tests__/unit/__snapshots__
 
 # macOS
 .DS_Store
