+++ __tests__/unit/BuildingHighlight.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { BuildingHighlight } from '../components/BuildingHighlight';
+import { BuildingHighlight } from '../../components/BuildingHighlight';
 import renderer from 'react-test-renderer';
 
 describe("Building Highlight component", () => {
